@@ -16,6 +16,7 @@
     "lint": "npm run lint-js && npm run lint-css",
     "lint-css": "stylelint './src/**/*.css'",
     "lint-js": "eslint '**/*.js' --ignore-path .gitignore",
+    "lint-js-fix": "eslint '**/*.js' --fix --ignore-path .gitignore",
     "postinstall": "npm run build",
     "start": "cross-env NODE_ENV=production node --use_strict server/node-server.js",
     "pretest": "npm run clean coverage/ && npm run lint",
@@ -51,9 +52,9 @@
     "es5-shim": "^4.5.9",
     "es6-promise": "^3.2.1",
     "es6-shim": "^0.35.1",
-    "eslint": "^3.2.1",
-    "eslint-loader": "^1.4.1",
-    "eslint-plugin-react": "^6.0.0",
+    "eslint": "^3.19.0",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-react": "^6.10.3",
     "file-loader": "^0.9.0",
     "history": "^3.0.0",
     "html-webpack-plugin": "^2.9.0",
@@ -91,7 +92,6 @@
     "redux-localstorage": "^0.4.0",
     "redux-logger": "^2.1.4",
     "redux-thunk": "^2.0.1",
-    "require-hacker": "^2.1.2",
     "rimraf": "^2.4.3",
     "sinon": "^1.17.3",
     "style-loader": "^0.13.0",
